requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
google-ads==21.3.0
beautifulsoup4==4.12.2
pydantic==2.5.0